Fixed blog apis

+++ src/main/java/com/cisco/cmad/blogservice/impl/BlogManagerImpl.java
@@ -1,5 +1,6 @@
 package com.cisco.cmad.blogservice.impl;
 
+import java.util.Date;
 import java.util.List;
 
 import com.cisco.cmad.blogservice.api.Blog;
@@ -9,15 +10,18 @@ import com.cisco.cmad.blogservice.api.BlogNotFoundException;
 import com.cisco.cmad.blogservice.api.DuplicateBlogException;
 import com.cisco.cmad.blogservice.api.InvalidBlogException;
 import com.cisco.cmad.blogservice.api.NotAuthorizedException;
+import com.cisco.cmad.blogservice.api.User;
+import com.cisco.cmad.blogservice.api.UserManager;
 import com.cisco.cmad.blogservice.dao.api.BlogDAO;
 import com.cisco.cmad.blogservice.dao.jpa.JPABlogDAO;
 
 public class BlogManagerImpl implements BlogManager {
 
-	BlogDAO blogDAO = new JPABlogDAO();
+	 BlogDAO blogDAO = new JPABlogDAO();
+	 UserManager userManager = new UserManagerImpl();
 
 	@Override
-	public Blog createBlog(Blog blog) throws DuplicateBlogException, InvalidBlogException, BlogException {
+	public Blog createBlog(String userId, String tocken, Blog blog) throws DuplicateBlogException, InvalidBlogException, BlogException {
 		if (blog == null) {
 			throw new InvalidBlogException();
 		}
@@ -25,9 +29,15 @@ public class BlogManagerImpl implements BlogManager {
 				|| ((blog.getName() == null) || (blog.getName().trim().isEmpty()))) {
 			throw new InvalidBlogException();
 		}
-		
 		Blog createdBlog = null;
 		try {
+			User user = userManager.getUser(tocken, userId);
+			if (user == null) {
+				throw new NotAuthorizedException();
+			}
+			blog.setCreated(new Date());
+			blog.setLastModifed(new Date());
+			blog.setUser(user);
 			createdBlog = blogDAO.create(blog);
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -52,7 +62,7 @@ public class BlogManagerImpl implements BlogManager {
 	}
 
 	@Override
-	public Blog updateBlog(Blog blog)
+	public Blog updateBlog(String userId, String tocken, Blog blog)
 			throws BlogNotFoundException, NotAuthorizedException, InvalidBlogException, BlogException {
 		if (blog == null) {
 			throw new InvalidBlogException();
@@ -66,8 +76,16 @@ public class BlogManagerImpl implements BlogManager {
 			throw new BlogNotFoundException();
 		}
 		Blog updatedBlog = null;
+		User user = userManager.getUser(tocken, userId);
+		if (user == null) {
+			throw new NotAuthorizedException();
+		}
 		try {
-			updatedBlog = blogDAO.update(blog);
+			existingBlog.setLastModifed(new Date());
+			existingBlog.setData(blog.getData());
+			existingBlog.setName(blog.getName());
+			updatedBlog = blogDAO.update(existingBlog);
+			
 		} catch (Exception e) {
 			e.printStackTrace();
 			throw new BlogException();
@@ -77,11 +95,15 @@ public class BlogManagerImpl implements BlogManager {
 	}
 
 	@Override
-	public void deleteBlog(int blogId) throws BlogNotFoundException, NotAuthorizedException, BlogException {
+	public void deleteBlog(String userId, String tocken, int blogId) throws BlogNotFoundException, NotAuthorizedException, BlogException {
 		try {
 			Blog blog = blogDAO.get(blogId);
 			if(blog == null) {
 				throw new BlogNotFoundException();
+			}
+			User user = userManager.getUser(tocken, userId);
+			if (user == null) {
+				throw new NotAuthorizedException();
 			}
 			blogDAO.delete(blog);
 		} catch (Exception e) {
